+++ src/app/features/blog/blog.routes.ts
@@ -8,6 +8,10 @@ export const blogRoutes: Routes = [
     loadComponent: () => import('./pages/blog-index/blog-index.component').then(m => m.BlogIndexComponent),
   },
   {
+    path: `${PATH_NAMES.BLOG}/category/:category`,
+    loadComponent: () => import('./pages/blog-category/blog-category.component').then(m => m.BlogCategoryComponent),
+  },
+  {
     path: `${PATH_NAMES.BLOG}/:slug`,
     loadComponent: () => import('./pages/blog-detail/blog-detail.component').then(m => m.BlogDetailComponent),
   },